label_ocr_api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
httpx>=0.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
pydantic>=2.0.0
python-multipart>=0.0.6
